% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyRatings.R
\name{updateShinyRatings}
\alias{updateShinyRatings}
\title{Change the value of star ratings on the client}
\usage{
updateShinyRatings(session, inputId, disable = FALSE)
}
\arguments{
\item{session}{The session object passed to function in the server.}

\item{inputId}{The input slot that will be used to access the value of number of stars.}

\item{disable}{default (FALSE). Do you want to disable the input?}
}
\description{
Change the value of star ratings on the client
}
