% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyRatings.R
\name{shinyRatings}
\alias{shinyRatings}
\title{Include ratings in shiny UI}
\usage{
shinyRatings(inputId, label)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value of number of stars.}

\item{label}{Display label on top of ratings stars}
}
\value{
Ratings to be added to UI definition
}
\description{
Include ratings in shiny UI
}
